+++ src/main/java/net/chenlin/dp/modules/biz/member/entity/LastLoginDateEntity.java
@@ -20,6 +20,6 @@ public class LastLoginDateEntity implements Serializable {
 
     private String memberId;
 
-    private int org_id;
+    private Long org_id;
 
 }
